+++ controllers/index.js
@@ -1,0 +1,8 @@
+const { getStatistics } = require("./statistics");
+const { getAllTransactions, createTransactions } = require("./transactions");
+
+module.exports = {
+  getStatistics,
+  getAllTransactions,
+  createTransactions,
+};
