+++ helpers/statistic/amountByTransactionType.js
@@ -1,10 +1,10 @@
-function amountByTransactionType(arr, type) {
+const amountByTransactionType = (arr, type) => {
   return arr.reduce(
     (acc, { amount, isIncome }) =>
       isIncome === type ? (acc * 100 + amount * 100) / 100 : acc,
     0
   );
-}
+};
 
 module.exports = {
   amountByTransactionType,
