+++ helpers/transaction/addCategoryIdName.js
@@ -1,0 +1,10 @@
+const addCategoryIdName = (obj) => {
+  const newArr = [];
+  obj.map(({ id, name }) => {
+    newArr.push([id, name]);
+  });
+  return newArr;
+};
+module.exports = {
+  addCategoryIdName,
+};
