+++ helpers/transaction/countTheBalance.js
@@ -1,8 +1,8 @@
-function countTheBalance(isIncome, balance, amount) {
+const countTheBalance = (isIncome, balance, amount) => {
   return isIncome === true
     ? (balance * 100 + amount * 100) / 100
     : (balance * 100 - amount * 100) / 100;
-}
+};
 
 module.exports = {
   countTheBalance,
